--- src/core/components/radio/stories.tsx
@@ -1,12 +1,0 @@
-export default {
-	title: "Radio",
-}
-
-export * from "./stories/controlled"
-export * from "./stories/error"
-export * from "./stories/horizontal"
-export * from "./stories/supporting-text"
-export * from "./stories/supporting-text-only"
-export * from "./stories/ungrouped"
-export * from "./stories/unlabelled"
-export * from "./stories/vertical"
