+++ src/core/components/radio/stories/radio/unlabelled.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { Radio, radioDefault } from "../index"
+import { Radio, radioDefault } from "../../index"
 import { ThemeProvider } from "emotion-theming"
 
 const unlabelled = () => (
